Move Resolver/AbstractResolver into own module.
