+++ src/Stack/BuildPlan.hs
@@ -93,6 +93,7 @@ import           Stack.Types.Version
 import           Stack.Types.Config
 import           Stack.Types.Urls
 import           Stack.Types.Compiler
+import           Stack.Types.Resolver
 import           Stack.Types.StackT
 
 data BuildPlanException
