+++ src/Stack/Init.hs
@@ -48,6 +48,7 @@ import           Stack.Types.BuildPlan
 import           Stack.Types.Config
 import           Stack.Types.FlagName
 import           Stack.Types.PackageName
+import           Stack.Types.Resolver
 import           Stack.Types.StackT (StackM)
 import           Stack.Types.Version
 import qualified System.FilePath as FP
