+++ src/Stack/Options/ResolverParser.hs
@@ -5,8 +5,8 @@ import qualified Data.Text                         as T
 import           Options.Applicative
 import           Options.Applicative.Types         (readerAsk)
 import           Stack.Options.Utils
-import           Stack.Types.Config
 import           Stack.Types.Compiler
+import           Stack.Types.Resolver
 
 -- | Parser for the resolver
 abstractResolverOptsParser :: Bool -> Parser AbstractResolver
