+++ src/Stack/Solver.hs
@@ -74,6 +74,7 @@ import           Stack.Types.Config
 import           Stack.Types.FlagName
 import           Stack.Types.PackageIdentifier
 import           Stack.Types.PackageName
+import           Stack.Types.Resolver
 import           Stack.Types.StackT (StackM)
 import           Stack.Types.Version
 import qualified System.Directory as D
