+++ stack.cabal
@@ -165,6 +165,7 @@ library
                      Stack.Types.PackageIdentifier
                      Stack.Types.PackageIndex
                      Stack.Types.PackageName
+                     Stack.Types.Resolver
                      Stack.Types.Sig
                      Stack.Types.StackT
                      Stack.Types.TemplateName
